src/mod-4-multi-agents/crew-ai-agents/requirements.txt
crewai==0.165.1
crewai-tools==0.62.3
docling>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
supabase>=2.5.0
vecs>=0.4.0
pgvector>=0.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
openai>=1.35.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
tiktoken>=0.7.0
langfuse>=2.50.0
openlit>=1.0.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation>=0.45b0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jsonschema>=4.22.0
asyncio>=3.4.3
aiohttp>=3.9.0
aiofiles>=23.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
pytest-cov>=5.0.0
pytest-env>=1.1.0
faker>=25.0.0
factory-boy>=3.3.0
black>=24.4.0
ruff>=0.4.0
mypy>=1.10.0
pre-commit>=3.7.0
ipython>=8.24.0
jupyter>=1.0.0
rich>=13.7.0
python-dateutil>=2.9.0
pytz>=2024.1
httpx>=0.27.0
tenacity>=8.3.0
structlog>=24.2.0
click>=8.1.0
numpy>=1.26.0
numba>=0.59.0
uvloop>=0.19.0
redis>=5.0.0
diskcache>=5.6.0